requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==3.0.2
requests==2.33.1
openrouter==0.9.1
SQLAlchemy==2.0.49
faiss-cpu==1.13.2
sentence-transformers==5.4.1